+++ public/schema/cloud_allow_list.txt
@@ -80,3 +80,4 @@ reject
 reject_errored
 retry
 switch
+none
