+++ resources/docker/Dockerfile.cloud
@@ -1,3 +1,11 @@
+# Copyright 2024 Redpanda Data, Inc.
+#
+# Licensed as a Redpanda Enterprise file under the Redpanda Community
+# License (the "License"); you may not use this file except in compliance with
+# the License. You may obtain a copy of the License at
+#
+# https://github.com/redpanda-data/connect/blob/main/licenses/rcl.md
+
 FROM golang:1.22 AS build
 
 ENV CGO_ENABLED=0
